+++ pkg/hls/manager.go
