+++ pkg/hlsproxy/cache.go
