+++ pkg/hlsproxy/manager.go
