+++ pkg/hlsproxy/types.go
